Made Tawk work on load
