+++ components/Layout/Layout.tsx
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import CookieConsent from 'react-cookie-consent';
+import CookieConsent, { getCookieConsentValue } from 'react-cookie-consent';
 import Head from 'next/head';
+import Tawk from '../../lib/tawk';
 
 type Props = {
   children: React.ReactNode;
@@ -45,6 +46,13 @@ export default function Layout(props: Props) {
     return () => {
       window.removeEventListener('scroll', navLocation);
     };
+  });
+
+  // onload, check if we show chat bot
+  useEffect(() => {
+    if (getCookieConsentValue()) {
+      Tawk();
+    }
   });
 
   const mortgageLinks = [
@@ -268,18 +276,8 @@ export default function Layout(props: Props) {
           declineButtonStyle={{ background: '#000000', color: '#ffffff' }}
           buttonText="Allow cookies"
           buttonStyle={{ background: '#29E458', fontWeight: '600' }}
-          onAccept={(acceptedByScrolling) => {
-            var Tawk_API = Tawk_API || {},
-              Tawk_LoadStart = new Date();
-            (function () {
-              var s1 = document.createElement('script'),
-                s0 = document.getElementsByTagName('script')[0];
-              s1.async = true;
-              s1.src = 'https://embed.tawk.to/5e258b068e78b86ed8aa286b/default';
-              s1.charset = 'UTF-8';
-              s1.setAttribute('crossorigin', '*');
-              s0.parentNode.insertBefore(s1, s0);
-            })();
+          onAccept={() => {
+            Tawk();
           }}
         >
           We use{' '}
