+++ lib/tawk.ts
@@ -1,0 +1,14 @@
+const Tawk = () => {
+  var Tawk_API = Tawk_API || {};
+  (function () {
+    var s1 = document.createElement('script'),
+      s0 = document.getElementsByTagName('script')[0];
+    s1.async = true;
+    s1.src = 'https://embed.tawk.to/5e258b068e78b86ed8aa286b/default';
+    s1.charset = 'UTF-8';
+    s1.setAttribute('crossorigin', '*');
+    s0.parentNode.insertBefore(s1, s0);
+  })();
+};
+
+export default Tawk;
